IS: Move kafkaDTOs to domain-folder
