+++ src/main/kotlin/no/nav/syfo/aktivitetskrav/kafka/AktivitetskravVarselProducer.kt
@@ -1,6 +1,7 @@
 package no.nav.syfo.aktivitetskrav.kafka
 
 import no.nav.syfo.aktivitetskrav.domain.AktivitetskravVarsel
+import no.nav.syfo.aktivitetskrav.kafka.domain.KafkaAktivitetskravVarsel
 import no.nav.syfo.domain.PersonIdent
 import org.apache.kafka.clients.producer.KafkaProducer
 import org.apache.kafka.clients.producer.ProducerRecord
