+++ src/main/kotlin/no/nav/syfo/aktivitetskrav/kafka/AktivitetskravVurderingProducer.kt
@@ -3,6 +3,7 @@ package no.nav.syfo.aktivitetskrav.kafka
 import no.nav.syfo.aktivitetskrav.domain.Aktivitetskrav
 import no.nav.syfo.aktivitetskrav.domain.AktivitetskravStatus
 import no.nav.syfo.aktivitetskrav.domain.toKafkaAktivitetskravVurdering
+import no.nav.syfo.aktivitetskrav.kafka.domain.KafkaAktivitetskravVurdering
 import no.nav.syfo.application.kafka.KafkaEnvironment
 import no.nav.syfo.application.kafka.kafkaAivenProducerConfig
 import org.apache.kafka.clients.producer.KafkaProducer
