+++ src/main/kotlin/no/nav/syfo/aktivitetskrav/kafka/KafkaAktivitetskravVarselSerializer.kt
@@ -1,5 +1,6 @@
 package no.nav.syfo.aktivitetskrav.kafka
 
+import no.nav.syfo.aktivitetskrav.kafka.domain.KafkaAktivitetskravVarsel
 import no.nav.syfo.util.configuredJacksonMapper
 import org.apache.kafka.common.serialization.Serializer
