+++ src/main/kotlin/no/nav/syfo/aktivitetskrav/kafka/domain/KafkaAktivitetskravVarsel.kt
@@ -1,4 +1,4 @@
-package no.nav.syfo.aktivitetskrav.kafka
+package no.nav.syfo.aktivitetskrav.kafka.domain
 
 import no.nav.syfo.aktivitetskrav.api.DocumentComponentDTO
 import java.time.LocalDate
